Rename(#31): Properties 위치 수정
